+++ src/main/java/net/thedragonteam/armorplus/registry/ModItems.java
@@ -105,7 +105,6 @@ import net.thedragonteam.armorplus.items.bows.*;
 import net.thedragonteam.armorplus.items.consumables.RedstoneApple;
 import net.thedragonteam.armorplus.items.consumables.TheGiftOfTheGods;
 import net.thedragonteam.armorplus.items.dev.*;
-import net.thedragonteam.armorplus.items.energy.ItemTeslaEnergy;
 import net.thedragonteam.armorplus.items.energy.ItemTeslaPickaxe;
 import net.thedragonteam.armorplus.items.energy.ItemTeslaSword;
 import net.thedragonteam.armorplus.items.materials.*;
@@ -367,7 +366,6 @@ public class ModItems {
     public static BeamItem beamItem;
     public static MMDItem mmdItem;
 
-    public static ItemTeslaEnergy itemTeslaEnergy;
     public static ItemTeslaPickaxe itemTeslaPickaxe;
     public static ItemTeslaSword itemTeslaSword;
 
@@ -378,7 +376,6 @@ public class ModItems {
     }
 
     public static void init() {
-        itemTeslaEnergy = new ItemTeslaEnergy();
         itemTeslaPickaxe = new ItemTeslaPickaxe();
         itemTeslaSword = new ItemTeslaSword();
         mmdItem = new MMDItem();
@@ -532,7 +529,6 @@ public class ModItems {
 
     @SideOnly(Side.CLIENT)
     public static void initModels() {
-        itemTeslaEnergy.initModel();
         itemTeslaPickaxe.initModel();
         itemTeslaSword.initModel();
         mmdItem.initModel();