+++ src/main/java/net/thedragonteam/armorplus/registry/ModRecipes.java
@@ -106,6 +106,9 @@ public class ModRecipes {
 
         GameRegistry.addRecipe(new ItemStack(ModBlocks.lavaNetherBrick, 4), "XNX", "NLN", "XNX", 'L', Items.LAVA_BUCKET, 'N', Blocks.NETHER_BRICK);
 
+        GameRegistry.addRecipe(new ItemStack(ModItems.itemTeslaSword, 1), "XTX", "XTX", "XSX", 'T', ModItems.steelIngot, 'S', Items.STICK);
+        GameRegistry.addRecipe(new ItemStack(ModItems.itemTeslaPickaxe, 1), "TTT", "XSX", "XSX", 'T', ModItems.steelIngot, 'S', Items.STICK);
+
         if (recipes == 0 && enableArrowRecipes) {
             GameRegistry.addRecipe(new ItemStack(ModItems.coalArrow, 2), "CCC", "CAC", "CCC", 'C', Items.COAL, 'A', Items.ARROW);
             GameRegistry.addRecipe(new ItemStack(ModItems.lapisArrow, 2), "LLL", "LAL", "LLL", 'L', lapisLAZULI, 'A', Items.ARROW);